wip: refactor on rmid management
