+++ Makefile
@@ -1,10 +1,21 @@
-KERNELDIR ?= /lib/modules/$(shell uname -r)/build
+KERNELDIR   ?= /lib/modules/$(shell uname -r)/build
 MODULE_NAME := hrperf
 
 obj-m := $(MODULE_NAME).o
-$(MODULE_NAME)-objs := ./src/buffer.o ./src/log.o ./src/hrperf.o ./src/cpucounters.o ./src/mmio.o ./src/uncore_pmu_discovery.o ./src/uncore_pmu.o ./src/mbm/rmid.o
+$(MODULE_NAME)-objs := \
+	./src/buffer.o \
+	./src/log.o \
+	./src/hrperf.o \
+	./src/cpucounters.o \
+	./src/mmio.o \
+	./src/uncore_pmu_discovery.o \
+	./src/uncore_pmu.o \
+	./src/mbm/rmid.o \
+	./src/mbm/mbm.o
 
 ccflags-y := -I$(PWD)/include -Wall -g
+
+.PHONY: all submake install clean
 
 all: submake
 	$(MAKE) -C $(KERNELDIR) M=$(PWD) modules
@@ -23,5 +34,3 @@ clean:
 	$(MAKE) -C ./src/io clean
 	rm -rf ./install
 	rm -f .*.cmd *.mod.c Module.symvers modules.order
-
-.PHONY: all submake install clean
