+++ include/mbm/mbm.h
@@ -1,0 +1,8 @@
+#ifndef _MBM_H_
+#define _MBM_H_
+
+int mbm_init(void);
+void mbm_init_cap(void);
+int mbm_deinit(void);
+
+#endif /* _MBM_H_ */
